refactor: use df2_simd::D2<4> as the only default
